Move model service into agent and remove models package
