--- packages/models/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "@amby/models",
-	"version": "0.1.0",
-	"private": true,
-	"type": "module",
-	"exports": {
-		".": "./src/index.ts"
-	},
-	"scripts": {
-		"typecheck": "tsc --noEmit",
-		"build": "bun build ./src/index.ts --outdir dist --target bun"
-	},
-	"dependencies": {
-		"@ai-sdk/openai": "^1.3.0",
-		"@amby/env": "workspace:*",
-		"@openrouter/ai-sdk-provider": "^2.3.1"
-	},
-	"peerDependencies": {
-		"ai": "^6.0.116",
-		"effect": "^3.19.19"
-	}
-}
