--- packages/models/src/errors.ts
@@ -1,6 +1,0 @@
-import { Data } from "effect"
-
-export class ModelError extends Data.TaggedError("ModelError")<{
-	readonly message: string
-	readonly cause?: unknown
-}> {}
