--- packages/models/src/providers/index.ts
@@ -1,2 +1,0 @@
-export * from "./stt"
-export * from "./tts"
