--- packages/models/tsconfig.json
@@ -1,4 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"include": ["src"]
-}
